+++ public/index.html
@@ -41,7 +41,7 @@
 
     <meta
       property="og:image"
-      content="%PUBLIC_URL%/images/DataAnalyzerScreenshot-Main-Zoomed.png"
+      content="%PUBLIC_URL%/images/DataAnalyzer-EventData-Screenshot.png"
     />
     <meta property="og:image:width" content="1200" />
     <meta property="og:image:height" content="630" />
