+++ src/components/DemoDataMenu/index.tsx
@@ -1,0 +1,2 @@
+import { DemoDataMenu } from './DemoDataMenu';
+export { DemoDataMenu };
