+++ src/components/SchemaTools/index.scss
@@ -52,6 +52,18 @@
   }
 }
 
+.roll-dice {
+  &.animated {
+    animation: roll-dice 1.5s steps(6) infinite;
+  }
+  display: inline-block;
+  // animation: roll-dice 1.5s steps(6) infinite;
+  background-image: url(../../../public/images/dice-sprite.svg);
+  width: 25px;
+  height: 25px;
+  background-size: cover;
+}
+
 // Credit: https://www.florin-pop.com/blog/2019/03/css-pulse-effect/
 // More reading: https://css-tricks.com/almanac/properties/a/animation/
 .pulse-error-icon {
@@ -77,5 +89,32 @@
   100% {
     transform: scale(0.65);
     box-shadow: 0 0 0 0 rgba(255, 82, 82, 0);
+  }
+}
+
+@keyframes roll-dice {
+  // 0% {
+  //   background-position: 0 -0px;
+  // }
+  // 16% {
+  //   background-position: 0 -100px;
+  // }
+  // 33% {
+  //   background-position: 0 -200px;
+  // }
+  // 49% {
+  //   background-position: 0 -300px;
+  // }
+  // 65% {
+  //   background-position: 0 -400px;
+  // }
+  // 81% {
+  //   background-position: 0 -500px;
+  // }
+  // 97% {
+  //   background-position: 0 -600px;
+  // }
+  100% {
+    background-position: -600px 0;
   }
 }
