+++ src/styles.scss
@@ -120,7 +120,7 @@ $breakpoint-xl: 1200p;
       }
       .resizable-editor {
         // flex-shrink: 1;
-        flex: 1 1 50%;
+        flex: 1 1 30vh;
       }
       .output-buttons-panel {
         flex: 0 0 70px;
@@ -131,7 +131,7 @@ $breakpoint-xl: 1200p;
     }
     &.results-loaded.stacked-view {
       .resizable-editor {
-        flex: 1 2 200px;
+        flex: 0 1 200px;
         min-height: 12vh;
       }
       .generated-code {
@@ -181,58 +181,6 @@ $breakpoint-xl: 1200p;
         height: 1.75rem;
         width: 1.75rem;
       }
-    }
-  }
-}
-
-.current-xs,
-.current-sm {
-  .demo-data-buttons {
-    order: 3;
-    flex-wrap: nowrap;
-    justify-content: flex-start;
-    overflow-x: auto;
-  }
-}
-.schema-name-input {
-  order: 1;
-}
-// Header Styles
-.demo-data-buttons {
-  order: 3;
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: flex-end;
-  align-items: flex-end;
-  // align-items: center;
-  position: relative;
-  z-index: -1;
-
-  .MuiChip-root {
-    margin: 0 1.5px 1.5px 0;
-    display: flex;
-    justify-content: flex-start;
-    svg {
-      margin-left: 2.5px;
-    }
-  }
-
-  .MuiChip-label {
-    flex-grow: 1;
-    text-align: center;
-  }
-
-  label {
-    margin-right: 3px;
-    display: flex;
-    // align-items: center;
-    justify-content: center;
-    margin: 0;
-    line-height: 24px;
-    // height: 24px;
-    font-weight: 500;
-    svg {
-      margin: 0 2.5px;
     }
   }
 }
