+++ tsconfig.json
@@ -2,8 +2,9 @@
   "extends": "./tsconfig.paths.json",
   "compilerOptions": {
     "baseUrl": "src",
-    "target": "ES2020",
+    "target": "ES2019",
     "lib": [
+      "ESNext.BigInt",
       "es2019",
       "dom",
       "dom.iterable",
@@ -30,6 +31,7 @@
     "importHelpers": true
   },
   "include": [
+    "**/__tests__/*.tsx",
     "src/**/*.ts",
     "src/**/*.tsx",
     "src/**/*.js",
